cast to same integer signedness for comparison

+++ SignalMessaging/profiles/OWSProfileManager.m
@@ -585,7 +585,7 @@ typedef void (^ProfileManagerFailureBlock)(NSError *error);
         uint8_t byte = (uint8_t)(0xff & byteValue);
         [groupId appendBytes:&byte length:1];
     }
-    if (groupId.length != kGroupIdLength) {
+    if (groupId.length != (NSUInteger)kGroupIdLength) {
         OWSFailDebug(@"Parsed group id has unexpected length: %@ (%lu)",
             groupId.hexadecimalString,
             (unsigned long)groupId.length);
